requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/base.in
#
argon2-cffi==20.1.0       # via -r requirements/base.in
asgiref==3.3.0            # via django
certifi==2020.6.20        # via elasticsearch, requests
cffi==1.14.3              # via argon2-cffi
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via python3-openid
django-allauth==0.42.0    # via -r requirements/base.in
django-appconf==1.0.4     # via django-compressor
django-compressor==2.4    # via -r requirements/base.in, django-libsass
django-controlled-vocabulary==0.10  # via django-ead
django-cors-headers==3.5.0  # via -r requirements/base.in
django-crispy-forms==1.9.2  # via -r requirements/base.in
django-ead==0.1.2         # via -r requirements/base.in
django-elasticsearch-dsl==7.1.4  # via -r requirements/base.in
django-environ==0.4.5     # via -r requirements/base.in
django-formtools==2.2     # via -r requirements/base.in
django-libsass==0.8       # via -r requirements/base.in
django-model-utils==4.0.0  # via -r requirements/base.in, django-ead
django-redis==4.12.1      # via -r requirements/base.in
django==3.0.10            # via -r requirements/base.in, django-allauth, django-appconf, django-controlled-vocabulary, django-cors-headers, django-formtools, django-model-utils, django-redis, djangorestframework
djangorestframework==3.11.1  # via -r requirements/base.in
elasticsearch-dsl==7.3.0  # via -r requirements/base.in, django-elasticsearch-dsl
elasticsearch==7.9.1      # via elasticsearch-dsl
hiredis==1.1.0            # via -r requirements/base.in
idna==2.10                # via requests
libsass==0.20.1           # via django-libsass
lxml==4.5.2               # via django-ead
oauthlib==3.1.0           # via requests-oauthlib
pillow==7.2.0             # via -r requirements/base.in
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via elasticsearch-dsl
python-slugify==4.0.1     # via -r requirements/base.in
python3-openid==3.2.0     # via django-allauth
pytz==2020.1              # via -r requirements/base.in, django
rcssmin==1.0.6            # via -r requirements/base.in, django-compressor
redis==3.5.3              # via -r requirements/base.in, django-redis
requests-oauthlib==1.3.0  # via django-allauth
requests==2.24.0          # via django-allauth, requests-oauthlib
rjsmin==1.1.0             # via django-compressor
six==1.15.0               # via argon2-cffi, django-compressor, django-elasticsearch-dsl, elasticsearch-dsl, libsass, python-dateutil
sqlparse==0.4.1           # via django
text-unidecode==1.3       # via python-slugify
urllib3==1.25.10          # via django-controlled-vocabulary, elasticsearch, requests
